Move create_component syscall into package
